Move worker files into separate worker folder t push
