+++ tests/api.v2/controllers/experimentController.test.js
@@ -4,7 +4,7 @@ const UserAccess = require('../../../src/api.v2/model/UserAccess');
 const { mockSqlClient, mockTrx } = require('../mocks/getMockSqlClient')();
 
 const getPipelineStatus = require('../../../src/api.v2/helpers/pipeline/getPipelineStatus');
-const getWorkerStatus = require('../../../src/api.v2/helpers/pipeline/getWorkerStatus');
+const getWorkerStatus = require('../../../src/api.v2/helpers/worker/getWorkerStatus');
 
 const experimentInstance = Experiment();
 const userAccessInstance = UserAccess();
@@ -25,7 +25,7 @@ jest.mock('../../../src/sql/sqlClient', () => ({
   get: jest.fn(() => mockSqlClient),
 }));
 jest.mock('../../../src/api.v2/helpers/pipeline/getPipelineStatus');
-jest.mock('../../../src/api.v2/helpers/pipeline/getWorkerStatus');
+jest.mock('../../../src/api.v2/helpers/worker/getWorkerStatus');
 
 const getExperimentResponse = require('../mocks/data/getExperimentResponse.json');
 const getAllExperimentsResponse = require('../mocks/data/getAllExperimentsResponse.json');
